style(styles): style file
Created independentily styles.
