+++ src/screens/Home/index.tsx
@@ -35,7 +35,6 @@ const Home = () => {
   );
 };
 
-// define your styles
 const styles = StyleSheet.create({
   container: {
     flex: 1,
@@ -49,5 +48,4 @@ const styles = StyleSheet.create({
   },
 });
 
-//make this component available to the app
 export default Home;
